requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.26
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.22
tiktoken>=0.5.2
pydantic>=2.5.3
